MemoryServer/requirements.txt
# Core server runtime
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
pydantic==2.11.7

# Shared retrieval/runtime deps (used by multiple baselines)
numpy>=1.24.3
openai>=1.3.7
sentence-transformers>=2.2.2
chromadb>=0.4.22
tqdm>=4.66.0

# A-Mem baseline
rank_bm25>=0.2.2
nltk>=3.8.1
transformers>=4.36.2
scikit-learn>=1.3.2

# mem0 baseline
pytz>=2024.1
